Adds job to install qcs and run camayoc against it
This job clones the quipucords/quipucords repo, installs necessary
system packages and installs other requirements via pip, and runs the
server for the duration of the job.

It does the same for quipucords/camayoc

It uses pytest to run camayoc functional tests against the qcs server,
and generates an xml report of the test results.

Closes #13